edited core data model, mdlanguagemodel, tests

+++ MyDictionary_App_Swift/Core/Models/MDLanguageModel.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct MDLanguageModel {
-    let id: UInt8
+    let id: Int16
     let name: String
     let translatedName: String
 }
@@ -24,7 +24,7 @@ extension MDLanguageModel: Decodable {
     
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = try container.decode(UInt8.self, forKey: .id)
+        self.id = try container.decode(Int16.self, forKey: .id)
         self.name = try container.decode(String.self, forKey: .name)
         self.translatedName = try container.decode(String.self, forKey: .translatedName)
     }
